requirements.txt
alembic==1.5.6
aniso8601==9.0.1
APScheduler==3.7.0
bcrypt==3.2.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.1.0
flask-marshmallow==0.14.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
mysqlclient==2.0.3
numpy==1.20.3
pycparser==2.20
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-engineio==4.1.0
python-socketio==5.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
tzlocal==2.1
urllib3==1.26.4
websockets==8.1
Werkzeug==1.0.1
